gg-communicator-py/requirements.txt
awsiotsdk==1.22.0
boto3==1.35.63